+++ app/components/pageComponentList.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import { useRouter } from 'next/navigation';
-import { FiSearch, FiFilter } from 'react-icons/fi';
+import { FiFilter } from 'react-icons/fi';
 import { AiOutlineSortAscending } from 'react-icons/ai';
-import { IoIosArrowDown } from "react-icons/io";
-import { FaLocationDot } from "react-icons/fa6";
-import { List, ListItem, ListItemText, ListItemAvatar, Avatar, Typography, Divider } from '@mui/material';
-import { useState } from 'react';
+import { List, ListItem, ListItemText, ListItemAvatar, Avatar, Typography } from '@mui/material';
 import Image from 'next/image';
 import Link from 'next/link';
 import { Box } from '@mui/system';
@@ -56,18 +53,7 @@ const GuideList: React.FC<{ guides: Guide[] }> = ({ guides }) => (
 );
 
   const PageComponentList: React.FC = () => {
-  const [searchTerm, setSearchTerm] = useState('');
   const router = useRouter();
-
-  const handleChangeLocation = () => {
-    // Implementar lógica para mudar a localização
-  };
-
-  // Função para lidar com a pesquisa
-  const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const term = event.target.value;
-    // Implemente a lógica para buscar o termo na página
-  };
 
   // Dados fictícios dos guias
   const guides: Guide[] = [
@@ -102,52 +88,6 @@ const GuideList: React.FC<{ guides: Guide[] }> = ({ guides }) => (
       margin: '0 auto', // Centralizar horizontalmente
       padding: '0 15px' // Adicionar margem fixa à esquerda e à direita
     }}>
-      {/* Localização atual */}
-      <div style={{ display: 'flex', alignItems: 'center' }}>
-        <div style={{ display: 'flex', alignItems: 'center' }}>
-          <FaLocationDot style={{ color: '#667358' }}/>
-          <span style={{ marginLeft: '10px' }}>Localização Atual</span>
-        </div>
-        <button 
-          onClick={handleChangeLocation} 
-          style={{ 
-            marginLeft: '10px',
-            backgroundColor: 'transparent',
-            border: 'none',
-            cursor: 'pointer'
-          }}>
-          <IoIosArrowDown />
-        </button>
-      </div>
-  
-      {/* Campo de busca */}
-      <div style={{ position: 'relative', marginTop: '10px'}}>
-        <input 
-          type="text" 
-          placeholder="Buscar parque, guia, trilha..." 
-          onChange={handleSearch} 
-          style={{ 
-            width: '100%', // Largura do campo de busca
-            padding: '5px', // Espaçamento interno
-            border: '1px solid #667358',
-            borderRadius: '4px'
-        }} />
-        <div style={{ position: 'absolute', right: '10px', top: '50%', transform: 'translateY(-50%)' }}>
-          <FiSearch />
-        </div>
-      </div>
-        
-      {/* Título da página */}
-      <h1 style={{ 
-        position: 'absolute', 
-        top: '15vh', 
-        left: '50%', 
-        transform: 'translate(-50%, -50%)',
-        textAlign: 'center',
-        textTransform: 'uppercase', 
-        fontSize: '18px', 
-        fontWeight: 'bold', 
-        color: '#667358'}}>Guias Credênciados</h1>
       
       {/* Ícone de filtro e ordenação */}
       <div style={{ position: 'relative', top: '10vh', display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
@@ -162,8 +102,6 @@ const GuideList: React.FC<{ guides: Guide[] }> = ({ guides }) => (
 
     </div>
   );
-  
-  
 };
 
 export default PageComponentList;
